refactor: file rename and dependencies update
