+++ src/routes/dashboard/budget/budget-table/budget-table.tsx
@@ -4,7 +4,7 @@ import { SubmitHandler, useFormContext } from 'react-hook-form';
 import { useBudgetTables } from '@/hooks/use-budget-tables';
 
 import { ButtonDelete, ButtonSecondary } from '@/components/button';
-import Modal from '@/components/modal';
+import Modal from '@/components/modal/modal';
 import MoreMenu from '@/components/more-menu';
 import { EditBudgetTableForm } from '@/routes/dashboard/budget/budget-table/budget-table-form';
 import { deleteBudgetTable, editBudgetTable } from '@/routes/dashboard/budget/budget-table/budget-table.api';
